Update _config.yaml separated login-registeration app
