+++ COUNCELLING_BOT/BOT/admin.py
@@ -1,6 +1,2 @@
 from django.contrib import admin
-from .models import User
 
-@admin.register(User)
-class UserAdmin(admin.ModelAdmin):
-    list_display = ('id', 'name','email','password')
