--- COUNCELLING_BOT/BOT/models.py
@@ -1,13 +1,0 @@
-from django.db import models
-from django.contrib.auth.models import User
-from PIL import Image
-
-
-# Extending User Model Using a One-To-One Link
-class User(models.Model):
-    name = models.CharField(max_length=70)
-    email = models.EmailField(max_length=100)
-    password = models.CharField(max_length=100)
-
-    def __str__(self):
-        return self.name
