+++ COUNCELLING_BOT/BOT/urls.py
@@ -1,11 +1,3 @@
 from django.urls import path
 
-from . import views
 
-urlpatterns = [
-    path('', views.home, name='home'),  # Root URL points to the home view
-    path('register/', views.register, name='register'),
-    path('login/', views.user_login, name='login'),
-    path('logout/', views.user_logout, name='logout'),
-    path('dashboard/', views.dashboard, name='dashboard'),
-]
