+++ COUNCELLING_BOT/COUNCELLING_BOT/urls.py
@@ -20,5 +20,5 @@ from django.urls import path, include
 
 urlpatterns = [
     path("admin/", admin.site.urls),
-    path('', include('BOT.urls'))
+    path('', include('LOGIN2BOT.urls')),
 ]
