+++ COUNCELLING_BOT/LOGIN2BOT/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+from .models import User
+
+@admin.register(User)
+class UserAdmin(admin.ModelAdmin):
+    list_display = ('id', 'name','email','password')
