+++ COUNCELLING_BOT/LOGIN2BOT/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class Login2BotConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "LOGIN2BOT"
