+++ COUNCELLING_BOT/LOGIN2BOT/forms.py
@@ -1,0 +1,12 @@
+from django import forms
+from .models import User
+
+class ProfileUser(forms.ModelForm):
+    class Meta:
+        model = User
+        fields = ['name','email','password']
+        widgets = {
+            'name':forms.TextInput(attrs={'class': 'form-control'}),
+            'email':forms.EmailInput(attrs={'class': 'form-control'}),
+            'password':forms.PasswordInput(render_value=True, attrs={'class': 'form-control'}),
+        }
